+++ src/modules/backlight.cpp
@@ -5,17 +5,18 @@
 
 // ────────────────────── グローバル変数 ──────────────────────
 BrightnessMode currentBrightnessMode = BrightnessMode::Day;
-uint32_t luxSampleBuffer[MEDIAN_BUFFER_SIZE] = {};
+uint16_t luxSampleBuffer[MEDIAN_BUFFER_SIZE] = {};
 int luxSampleIndex = 0;
 
 
 // ────────────────────── 輝度測定 ──────────────────────
-static uint32_t measureLuxWithoutBacklight()
+// バックライトを消して輝度を測定
+static uint16_t measureLuxWithoutBacklight()
 {
     uint8_t prevB = display.getBrightness();
     display.setBrightness(0);
     delayMicroseconds(500);
-    uint32_t lux = CoreS3.Ltr553.getAlsValue();
+    uint16_t lux = CoreS3.Ltr553.getAlsValue();
     display.setBrightness(prevB);
     return lux;
 }
@@ -31,15 +32,15 @@ void updateBacklightLevel()
         return;
     }
 
-    uint32_t lux = measureLuxWithoutBacklight();
+    uint16_t lux = measureLuxWithoutBacklight();
 
     luxSampleBuffer[luxSampleIndex] = lux;
     luxSampleIndex = (luxSampleIndex + 1) % MEDIAN_BUFFER_SIZE;
 
-    uint32_t sorted[MEDIAN_BUFFER_SIZE];
+    uint16_t sorted[MEDIAN_BUFFER_SIZE];
     memcpy(sorted, luxSampleBuffer, sizeof(sorted));
     std::nth_element(sorted, sorted + MEDIAN_BUFFER_SIZE / 2, sorted + MEDIAN_BUFFER_SIZE);
-    uint32_t medianLux = sorted[MEDIAN_BUFFER_SIZE / 2];
+    uint16_t medianLux = sorted[MEDIAN_BUFFER_SIZE / 2];
 
     BrightnessMode newMode =
         (medianLux >= LUX_THRESHOLD_DAY)  ? BrightnessMode::Day  :
